Add ShowInterfacesSwitchport module
Without this change, the only way to get at the switchport data is through the
configuration, which becomes a problem when you want all interfaces present on a
box due to the call-per-interface implementation of config parsing.  Here we
implement support for the 'show interfaces switchport' which has much of the
same data, but does not come from the configuration, and is quite a bit faster
due to being a single call.